few minor improvements, also refactored music and projects to extend a custom viewport-pagination-view class to remove code duplication
