--- react-spa/src/react-components/RootPage/Music/Music.css
@@ -1,26 +1,0 @@
-.music {
-    min-height: 100vh;
-    max-height: 100vh;
-    height: 100vh;
-}
-
-.music > * {
-    transition: transform 1s ease;
-}
-
-.music .viewport {
-    min-height: 100vh;
-    height: 100vh;
-    width: 100%;
-    display: flex;
-    align-items: center;
-}
-
-.music .viewport > *{
-    width: 100%;
-}
-
-.music .scroll-down {
-    padding-top: 100px;
-    text-align: center;
-}
