+++ react-spa/src/react-components/RootPage/RootPage.css
@@ -62,14 +62,9 @@
     user-select: none;
 }
 @media (max-width: 601px) {
-    .rootpage .container {
-        margin-top: 120px;
-    }
-
     .rootpage h1 {
        font-size: 4em;
     }
-
 }
 
 .anim-fade-enter, .anim-fade-enter .d1, .anim-fade-enter .d2, .anim-fade-enter .d3, .anim-fade-enter .d4, .anim-fade-enter .d5 {
